requirements.txt
django
djangorestframework
markdown
pillow
django-imagekit
pytest-django
